refactor(middleware): remove middleware from route file
